#74 Add a currentFile variable to each file
That variable contains the relative name of the current file. It can be used for example to link to the GitHub source file (to offer editing the documentation).

Because it was getting out of hand, I had to move back all the steps into the main application config. This is much simpler now, even though that means we can't dynamically add modules.