Read excluded directories from the `.gitignore` file of the repository
This allows users to have their git ignored files and directories automatically
ignored by Couscous. To enable this feature, they will have to add the special
`%gitignore%` entry to their exclude list.

Why do we use a special entry? To keep the actual behavior unchanged and avoid
unexpected generation errors.

The actual implementation is a bit naïve: it will take each entry line from the
`.gitignore` and add them to the exclude list of Couscous. So we might end up
adding files and directories. It's hard to test every `.gitignore` entry to see
if the are actual directories: paths might be relative or patterns, starting
from the repository's root directory or not.

Moreover, the implementation does not support complex pattern as git would.