Fix signal handling to avoid unexpected exceptions
When killing Couscous with CTRL+C (SIGINT - 2) user got an exception telling him
the server went down. This is expected an no exception should be raised in this
case.

Moreover, when the server exited correctly (after receiving SIGTERM - 15), an
error was displayed and an exception was raised. Again this behavior was wrong:
everything exited successfully and the errors were misleading. This is fixed.

Now, when Couscous receives either SIGINT or SIGTERM, it will try to kill the
server process. If everything exits properly, the user only gets an information
message an no exception is raised. If Couscous fails to stop the server it will
display an error message and throw an exception containing the server's error
output.